requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
ratelim==0.1.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
